requirements.txt
pandas
psycopg2-binary
chardet